+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+## [0.21.0] 2026-04-26
 ### Added
 - Mermaid diagram rendering for Markdown documents and chat responses.